# Use sync.Map as a read cache to avoid lock contention