--- a/internal/evaluator/builder.go
+++ b/internal/evaluator/builder.go
@@ -29,11 +29,9 @@ import (
 // Builder is a build-through cache of message evaluators keyed off the provided
 // descriptor.
 type Builder struct {
-	// TODO: (TCN-1708) based on benchmarks, about 50% of CPU time is spent obtaining a read
-	//  lock on this mutex. Ideally, this can be reworked to be thread-safe while
-	//  minimizing the need to obtain a lock.
-	mtx         sync.RWMutex
+	mtx         sync.Mutex
 	cache       map[protoreflect.MessageDescriptor]*message
+	readCache   sync.Map // copy of cache optimized for concurrent loads.
 	env         *cel.Env
 	constraints constraints.Cache
 	resolver    StandardConstraintResolver
@@ -81,16 +79,16 @@ func (bldr *Builder) load(desc protoreflect.MessageDescriptor) MessageEvaluator 
 // descriptor, or lazily constructs a new one. This method is thread-safe via
 // locking.
 func (bldr *Builder) loadOrBuild(desc protoreflect.MessageDescriptor) MessageEvaluator {
-	bldr.mtx.RLock()
-	if eval, ok := bldr.cache[desc]; ok {
-		bldr.mtx.RUnlock()
-		return eval
+	// Fast path uses readCache.
+	if eval, ok := bldr.readCache.Load(desc); ok {
+		return eval.(*message) //nolint:forcetypeassert
 	}
-	bldr.mtx.RUnlock()
 
 	bldr.mtx.Lock()
-	defer bldr.mtx.Unlock()
-	return bldr.build(desc)
+	msgEval := bldr.build(desc)
+	bldr.mtx.Unlock()
+	bldr.readCache.Store(desc, msgEval)
+	return msgEval
 }
 
 func (bldr *Builder) build(desc protoreflect.MessageDescriptor) *message {