Add a readCache to provide a copy of the cache optimized for concurrent loads. Reduces locking on lazy loading descriptors.

Fixes TCN-1708. 

```
benchstat current.txt new.txt
goos: darwin
goarch: arm64
pkg: github.com/bufbuild/protovalidate-go
                               │  current.txt  │                new.txt                │
                               │    sec/op     │    sec/op     vs base                 │
Validator/ColdStart-8             222.0µ ± ∞ ¹   216.1µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Validator/Lazy/Valid-8            150.9n ± ∞ ¹   118.4n ± ∞ ¹        ~ (p=1.000 n=1) ²
Validator/Lazy/Invalid-8          409.6n ± ∞ ¹   388.5n ± ∞ ¹        ~ (p=1.000 n=1) ²
Validator/Lazy/FailFast-8        143.30n ± ∞ ¹   98.49n ± ∞ ¹        ~ (p=1.000 n=1) ²
Validator/PreWarmed/Valid-8      108.30n ± ∞ ¹   98.44n ± ∞ ¹        ~ (p=1.000 n=1) ²
Validator/PreWarmed/Invalid-8     391.7n ± ∞ ¹   388.2n ± ∞ ¹        ~ (p=1.000 n=1) ²
Validator/PreWarmed/FailFast-8    98.66n ± ∞ ¹   95.91n ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                           503.7n         447.4n        -11.18%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                               │  current.txt  │                new.txt                │
                               │     B/op      │     B/op       vs base                │
Validator/ColdStart-8            421.6Ki ± ∞ ¹   421.8Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
Validator/Lazy/Valid-8             0.000 ± ∞ ¹     0.000 ± ∞ ¹       ~ (p=1.000 n=1) ³
Validator/Lazy/Invalid-8           649.0 ± ∞ ¹     649.0 ± ∞ ¹       ~ (p=1.000 n=1) ³
Validator/Lazy/FailFast-8          168.0 ± ∞ ¹     168.0 ± ∞ ¹       ~ (p=1.000 n=1) ³
Validator/PreWarmed/Valid-8        0.000 ± ∞ ¹     0.000 ± ∞ ¹       ~ (p=1.000 n=1) ³
Validator/PreWarmed/Invalid-8      649.0 ± ∞ ¹     649.0 ± ∞ ¹       ~ (p=1.000 n=1) ³
Validator/PreWarmed/FailFast-8     168.0 ± ∞ ¹     168.0 ± ∞ ¹       ~ (p=1.000 n=1) ³
geomean                                      ⁴                  +0.00%               ⁴
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
³ all samples are equal
⁴ summaries must be >0 to compute geomean

                               │ current.txt  │               new.txt                │
                               │  allocs/op   │  allocs/op    vs base                │
Validator/ColdStart-8            7.509k ± ∞ ¹   7.514k ± ∞ ¹       ~ (p=1.000 n=1) ²
Validator/Lazy/Valid-8            0.000 ± ∞ ¹    0.000 ± ∞ ¹       ~ (p=1.000 n=1) ³
Validator/Lazy/Invalid-8          15.00 ± ∞ ¹    15.00 ± ∞ ¹       ~ (p=1.000 n=1) ³
Validator/Lazy/FailFast-8         3.000 ± ∞ ¹    3.000 ± ∞ ¹       ~ (p=1.000 n=1) ³
Validator/PreWarmed/Valid-8       0.000 ± ∞ ¹    0.000 ± ∞ ¹       ~ (p=1.000 n=1) ³
Validator/PreWarmed/Invalid-8     15.00 ± ∞ ¹    15.00 ± ∞ ¹       ~ (p=1.000 n=1) ³
Validator/PreWarmed/FailFast-8    3.000 ± ∞ ¹    3.000 ± ∞ ¹       ~ (p=1.000 n=1) ³
geomean                                     ⁴                 +0.01%               ⁴
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
³ all samples are equal
⁴ summaries must be >0 to compute geomean
```